backend/requirements.txt
Django==4.0.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
python-json-logger==2.0.2